tests/requirements-test.txt
pytest==7.3.1
pytest-flask==1.2.0
beautifulsoup4==4.12.2
coverage==7.2.7
pytest-cov==4.1.0
